Remove warnings
Most of them were signed and unsigned comparison

+++ LeGO-LOAM/include/lego_loam/nanoflann_pcl.h
@@ -167,7 +167,7 @@ inline int KdTreeFLANN<PointT>::radiusSearch(
   }
   k_indices.resize(nFound);
   k_sqr_distances.resize(nFound);
-  for (int i = 0; i < nFound; i++) {
+  for (size_t i = 0; i < nFound; i++) {
     k_indices[i] = indices_dist[i].first;
     k_sqr_distances[i] = indices_dist[i].second;
   }
